securevault-backend/requirements.txt
alembic==1.16.2
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
